fix: Функции разбиты по единицам трансляции
